% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{colors2d}
\alias{colors2d}
\title{Map values to a 2D legend interpolated from 4 corner colors.}
\usage{
colors2d(data, colors = c("yellow", "green", "blue", "magenta"),
  xtrans = c("none", "log", "rank"), ytrans = c("none", "log", "rank"))
}
\arguments{
\item{data}{Matrix or data frame with 2 numeric columns; they will map to x
and y.}

\item{colors}{Vector of 4 corner colors to interpolate, clockwise from upper
right.}
}
\value{
Vector of color values.
}
\description{
This function returns a color value for each row of the 2-column dataset
supplied, based on a 2D color palette interpolated from 4 corner colors.
}
