% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors3d-package.R
\docType{package}
\name{colors3d-package}
\alias{colors3d-package}
\alias{_PACKAGE}
\title{colors3d: Generate 2D and 3D Color Palettes}
\description{
Generate multivariate color palettes to represent two-dimensional or three-dimensional data in graphics (in contrast to standard color palettes that represent just one variable). You tell 'colors3d' how to map color space onto your data, and it gives you a color for each data point. You can then use these colors to make plots in base 'R', 'ggplot2', or other graphics frameworks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matthewkling/colors3d}
  \item Report bugs at \url{https://github.com/matthewkling/colors3d/issues}
}

}
\author{
\strong{Maintainer}: Matthew Kling \email{mattkling@berkeley.edu} [copyright holder]

}
\keyword{internal}
