% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colomap.R
\name{colors3d}
\alias{colors3d}
\title{Get colors for a 3D dataset.}
\usage{
colors3d(data, trans = "fit", order = 1, inversion = 1, opacity = NULL)
}
\arguments{
\item{data}{Matrix or data frame with 3 numeric columns.}

\item{trans}{Either "fit" (defaut, histogram is rescaled) or "ecdf"
(histogram is flattened).}

\item{order}{Integer from 1 to 6, each denoting a unique permutation of
variables-to-color band mapping.}

\item{inversion}{Integer from 1 to 8, each denoting a unique combination of
variables to invert.}

\item{opacity}{Not currently used.}
}
\value{
Vector of color values.
}
\description{
This function returns a color value for each row of the 3-column dataset
supplied, by transforming the input data and using it as RGB values.
}

