% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colomap.R
\name{colorwheel2d}
\alias{colorwheel2d}
\title{Map values to a 2D colorwheel legend.}
\usage{
colorwheel2d(data, colors = c("black", "yellow", "green", "blue", "magenta"),
  origin = NULL, xyratio = NULL)
}
\arguments{
\item{data}{Matrix or data frame with 2 numeric columns; they will map to x
and y.}

\item{colors}{Vector of colors to interpolate: center followed by periphery
counterclockwise from 3 o'clock.}
}
\value{
Vector of color values.
}
\description{
This function returns a color value for each row of the 2-column dataset
supplied, based on a 2D color palette defined by a center color and a series
of peripheral colors.
}

