% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{polarize}
\alias{polarize}
\title{Internal function converting x-y do distance-angle.}
\usage{
polarize(data, xyratio, xorigin = 0, yorigin = 0)
}
\arguments{
\item{data}{Matrix or data frame with 2 numeric columns representing x and y.}

\item{xyratio}{Single number indicating unit ratio in x vs y direction.}

\item{xorigin, yorigin}{Numbers indicating center of polarization.}
}
\value{
2-column matrix of distances and angles.
}
\description{
Internal function converting x-y do distance-angle.
}
