% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colomap.R
\name{colors2d}
\alias{colors2d}
\title{Map values to a 2D legend interpolated from 4 corner colors.}
\usage{
colors2d(data, colors = c("yellow", "green", "blue", "magenta"))
}
\arguments{
\item{data}{Matrix or data frame with 3 numeric columns.}

\item{colors}{Vector of 4 corner colors to interpolate, clockwise from upper
right.}
}
\value{
Vector of color values.
}
\description{
This function returns a color value for each row of the 2-column dataset
supplied, based on a 2D color palette interpolated from 4 corner colors.
}

