% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{distant_colors}
\alias{distant_colors}
\title{Palettes of dissimilar colors in RGB space.}
\usage{
distant_colors(n)
}
\arguments{
\item{n}{Number of colors (integer).}
}
\value{
Vector of color values.
}
\description{
Many standard palette generators use only a slice of color space, which can
cause a lack of differentiability in palettes used to visualize categorical
factors with many levels. This function attempts to overcome this by
generating colors using clustering in 3D RGB space.
}

